requirements.txt
Flask==3.1.2
google-ads==28.0.0
python-dotenv==1.0.0
Werkzeug==3.1.2
psycopg2-binary==2.9.9
pandas==2.1.4

